backend/requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==13.1

# ─── Database ─────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
alembic==1.14.0

# ─── Cache / Queue ────────────────────────────────────────────────────────
redis==5.2.1

# ─── HTTP / Ingestion ─────────────────────────────────────────────────────
httpx==0.28.0
feedparser==6.0.11
playwright==1.49.0
asyncpraw==7.8.1

# ─── AI / Embeddings ──────────────────────────────────────────────────────
openai==1.57.4
google-generativeai==0.8.3

# ─── ML / Clustering ──────────────────────────────────────────────────────
numpy==2.2.0
scikit-learn==1.6.0
hdbscan==0.8.40

# ─── Config ───────────────────────────────────────────────────────────────
pydantic-settings==2.7.0
python-dotenv==1.0.1
